[chore] SearchService 구현을 위한 아키텍처 초기 구성 #27 - [rename] ArticleApplicationService -> ArticleCommandService: CUD 작업은 여기에서만
