+++ src/main/kotlin/dev/codeswamp/core/article/application/dto/ArticleIndexDTO.kt
@@ -1,0 +1,4 @@
+package dev.codeswamp.core.article.application.dto
+
+class ArticleIndexDTO {
+}
