+++ src/main/kotlin/dev/codeswamp/core/article/application/dto/ArticleSearchDTO.kt
@@ -1,0 +1,4 @@
+package dev.codeswamp.core.article.application.dto
+
+class ArticleSearchDTO {
+}
