+++ src/main/kotlin/dev/codeswamp/core/article/application/service/ArticleCommandService.kt
@@ -1,10 +1,8 @@
 package dev.codeswamp.core.article.application.service
 
 import dev.codeswamp.core.article.presentation.dto.request.ArticleMetadataDto
-import dev.codeswamp.core.article.presentation.dto.response.ArticleReadResponseDto
-import dev.codeswamp.core.user.infrastructure.persistence.entity.UserEntity
 
-interface ArticleApplicationService {
+interface ArticleCommandService {
     fun create(userId: Long, metadata: ArticleMetadataDto, content: String)
 
     fun updateArticleMetadata(userId: Long, articleId: Long, metadata: ArticleMetadataDto)
@@ -12,9 +10,4 @@ interface ArticleApplicationService {
     fun updateArticleContent(userId: Long, articleId: Long, content: String)
 
     fun delete(userId: Long, articleId: Long)
-
-    fun findByArticleId(userId: Long?, articleId: Long): ArticleReadResponseDto
-
-    fun findAllByKeywords(userId: Long?, keywords: List<String>): List<ArticleReadResponseDto>
-
 }
