+++ src/main/kotlin/dev/codeswamp/core/article/application/support/MarkdownPreprocessor.kt
@@ -1,0 +1,5 @@
+package dev.codeswamp.core.article.application.support
+
+interface MarkdownPreprocessor {
+    fun preprocess(rawMarkdown: String): String
+}
