--- src/main/kotlin/dev/codeswamp/core/article/infrastructure/search/ArticleSearchRepository.kt
@@ -1,5 +1,0 @@
-package dev.codeswamp.core.article.infrastructure.search
-
-interface ArticleSearchRepository {
-    fun searchArticleIdsByKeywords(keywords: List<String>): List<Long>
-}
