+++ src/main/kotlin/dev/codeswamp/core/article/infrastructure/search/MeiliSearchIndexer.kt
@@ -1,0 +1,4 @@
+package dev.codeswamp.core.article.infrastructure.search
+
+class MeiliSearchIndexer {
+}
