+++ src/main/kotlin/dev/codeswamp/core/article/infrastructure/search/config/MeiliSearchConfig.kt
@@ -1,0 +1,4 @@
+package dev.codeswamp.core.article.infrastructure.search.config
+
+class MeiliSearchConfig {
+}
