+++ src/main/kotlin/dev/codeswamp/core/article/infrastructure/support/FlexMarkdownPreprocessor.kt
@@ -1,0 +1,4 @@
+package dev.codeswamp.core.article.infrastructure.support
+
+class FlexMarkdownPreprocessor {
+}
